rag-chatbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
gradio==4.19.0
numpy==1.26.3
pydantic==2.6.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.3.1
torch==2.1.2
